檔名修正、component分類
群組中1-4點修正
