+++ APP/Client/components/NavStack/BelieverOfferingStack.js
@@ -2,7 +2,8 @@ import React from 'react'
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 
-import BelieverHomePage from '../../screens/Believer/BelieverHomePage';
+import OfferingPage from '../../screens//Believer/OfferingPage';
+import OfferingsByTemple from '../../screens/Believer/OfferingsByTemple';
 import TemplesByOffering from '../../screens/Believer/TemplesByOffering';
 import OrderConfirmationPage from '../../screens/Believer/OrderConfirmationPage';
 import OrderSuccess from '../../screens/Believer/OrderSuccess';
@@ -13,20 +14,25 @@ import ProfileManagement from '../../screens/Believer/ProfileManagement';
 
 const Stack = createNativeStackNavigator();
 
-function BelieverHomeStack(){
+function BelieverOfferingStack(){
   return (
     <Stack.Navigator
         screenOption = {{ headerShown: false }}
-        initialRouteName='BelieverHomePage'
+        initialRouteName='OfferingPage'
     >
       <Stack.Screen
-        name = 'BelieverHomePage'
-        component = { BelieverHomePage }
+        name = 'OfferingPage'
+        component = { OfferingPage }
         options = {{ headerShown: false }}
       />
       <Stack.Screen
         name = 'TemplesByOffering'
         component = { TemplesByOffering }
+        options = {{ headerShown: false }}
+      />
+      <Stack.Screen
+        name = 'OfferingsByTemple'
+        component = { OfferingsByTemple }
         options = {{ headerShown: false }}
       />
       <Stack.Screen
@@ -39,24 +45,9 @@ function BelieverHomeStack(){
         component = { OrderSuccess }
         options = {{ headerShown: false }}
       />
-      <Stack.Screen
-        name = 'OrderHistoryPage'
-        component = { OrderHistoryPage }
-        options = {{ headerShown: false }}
-      />
-      <Stack.Screen
-        name = 'SavedTemples'
-        component = { SavedTemples }
-        options = {{ headerShown: false }}
-      />
-      <Stack.Screen
-        name = 'ProfileManagement'
-        component = { ProfileManagement }
-        options = {{ headerShown: false }}
-      />
       
     </Stack.Navigator>
   )
 }
 
-export default BelieverHomeStack;
+export default BelieverOfferingStack;
