+++ APP/Client/screens/Believer/BelieverHomePage.js
@@ -5,14 +5,14 @@ import { Image } from "expo-image";
 import { useNavigation } from "@react-navigation/native";
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 
-import TempleDistance from "../../components/TempleDistance";
-import AddressOverlay from "../../components/AddressOverlay";
+import TempleDistance from "../../components/Believer/TempleDistance";
+import AddressOverlay from "../../components/Believer/AddressOverlay";
 
 import * as Location from 'expo-location';
 
 const { width } = Dimensions.get('window');
 
-const HomePage = () => {
+const BelieverHomePage = () => {
   const [locationIconVisible, setLocationIconVisible] = useState(false);
   const [text1Visible, setText1Visible] = useState(false);
   const [searchText, setSearchText] = useState("");
@@ -123,7 +123,7 @@ const HomePage = () => {
               distance={item.distance}
               date1={item.date1}
               date2={item.date2}
-              onPress={() => navigation.navigate("HomePage1")}
+              onPress={() => navigation.navigate("OfferingsByTemple")}
             />
           )}
           keyExtractor={(item) => item.id}
@@ -146,8 +146,6 @@ const HomePage = () => {
         </Modal>
 
 
-        {/* Footer */}  
-        {/* <Footer /> */}
       </View>
     </SafeAreaProvider>
   );
@@ -209,4 +207,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default HomePage;
+export default BelieverHomePage;
