--- APP/Client/screens/Welfare/Welfare_UserPage.js
@@ -1,24 +1,0 @@
-import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
-
-const Welfare_UserPage = () => {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.text}>This is the User Screen</Text>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  text: {
-    fontSize: 20,
-    color: 'black',
-  },
-});
-
-export default Welfare_UserPage;
